requirements.txt
playwright
httpx
psutil
google-genai
langfuse